Refine the design of ticket button as per Zeplin.

+++ app/src/main/java/com/routesme/vehicles/data/model/CarrierInformationModel.kt
@@ -4,7 +4,7 @@ class CarrierInformationModel {
 
     data class CarrierInformationModel(val routeNumber: String? = null, val destination: String? = null, val tickets: List<Ticket>, val included: Included, val message: String? = null, val status: Boolean = false, val statusCode: Int = -999)
 
-    data class Ticket(val ticket_id: String? = null, val amount: Double? = null, val first_station: String? = null, val currency_id: String? = null)
+    data class Ticket(val ticket_id: String? = null, val amount: Int? = null, val first_station: String? = null, val currency_id: String? = null)
     data class Included (val currencies: List<Currency>)
     data class Currency (val currency_id: String? = null, val code: String? = null, val symbol: String? = null)
